requirements.txt
Django==3.2.12
djangonepal==0.0.6
djangorestframework==3.14.0
requests==2.25.1
